+++ sdk-node/src/resources/subscriptions/index.ts
@@ -13,6 +13,7 @@ export type {
   SubscriptionInvoiceRef,
   SubscriptionItem,
   SubscriptionListParams,
+  SubscriptionManageUrl,
   SubscriptionProration,
   SubscriptionRetrieveParams,
   SubscriptionStatus,